feat: support explain plan statement.
